src/rag/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
spacy==3.5.3
nltk==3.8.1
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
pytest==7.3.1
requests==2.30.0
loguru==0.7.0
pyyaml==6.0
langdetect==1.0.9
